Add verbosity levels and logging output

+++ main.py
@@ -17,11 +17,14 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 # Or use this link: https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
 
+import argparse
 import hashlib
+import logging
 import os
 from datetime import datetime
 from os import path
-import argparse
+from sys import exit
+
 import requests
 
 # Get arguments
@@ -30,21 +33,32 @@ parser.add_argument('output_dir', type=str, nargs='?', default="~/Picture/NASA/"
                     help='Output dir for download script.')
 parser.add_argument('--url', '-u', type=str, nargs='?', default="https://apod.nasa.gov/apod/",
                     help='Custom URL to download from')
+parser.add_argument('--verbose', '-v', action='count', default=0,
+                    help="Increase verbosity - each statement increases by one")
 args = parser.parse_args()
+
+# Configure logging
+loglevel = [logging.CRITICAL, logging.ERROR, logging.WARNING, logging.INFO, logging.DEBUG]
+logging.basicConfig(level=loglevel[args.verbose if args.verbose <= 4 else 4],
+                    format='%(message)s')
 
 # Download image from NASA's APOD
 base_url = "https://apod.nasa.gov/apod/"
+logging.debug(f"Base URL: {base_url}")
 
 # Destination directory where files are stored
 destination_dir = path.expanduser(args.output_dir)
+logging.debug(f"Destination base path: {destination_dir}")
 
 # Location of the latest file (the last image is symlinked here at the end)
 latest_file = path.join(destination_dir, "latest")
+logging.debug(f"'latest' file located in: {latest_file}")
 
 # Download APOD webpage
 r = requests.get(args.url)
-if r.status_code == 404:
-    print(f"Page not found: {args.url}")
+if r.status_code != 200:
+    logging.info(f"Unable to download page: {args.url}")
+    exit(1)
 
 date = None
 
@@ -57,8 +71,10 @@ for line in r.text.split("\n"):
             date = line.split(':')[1]
             date = date.split('-')[0]
             date = datetime.strptime(date.strip(), "%Y %B %d")
+            logging.debug(f"Date found and set to: {date}")
         except Exception as e:
             date = datetime.now()
+            logging.debug(f"Did not find date in title - setting date to: {date}")
 
     # Download the image
     if "image/" in line.lower() and "href" in line.lower():
@@ -66,10 +82,12 @@ for line in r.text.split("\n"):
         # Set date if it's none
         if date is None:
             date = datetime.now()
+            logging.debug(f"No date available - using today: {date}")
 
         # Fetch the image url and create image link
         link = line.split('"')[1]
         image_link = f"{base_url}{link}"
+        logging.debug(f"Image link: {image_link}")
 
         # Output file
         destination_file = os.path.join(
@@ -79,13 +97,21 @@ for line in r.text.split("\n"):
                 path.basename(image_link)
             ])
         )
+        logging.debug(f"Destination file: {destination_file}")
+
+        # Image name
+        image_name = path.basename(destination_file)
+        logging.debug(f"Image file: {destination_file}")
 
         # Create directory if it does not exists
         if not path.isdir(path.dirname(destination_file)):
             os.makedirs(path.dirname(destination_file), exist_ok=True)
+            logging.info(f"Destination folder did not exist, creating {path.dirname(destination_file)}")
 
         # No need to do more work if file exists
-        if not path.isfile(destination_file):
+        if path.isfile(destination_file):
+            logging.error(f"The image '{image_name}' has already been downloaded previously")
+        else:
 
             # Grab the MD5 of the latest file
             if path.islink(latest_file):
@@ -93,33 +119,47 @@ for line in r.text.split("\n"):
                     with open(latest_file, "rb") as f:
                         last_image = f.read()
                     last_md5 = hashlib.md5(last_image).hexdigest()
+                    logging.info(f"md5sum of latest image: {last_md5}")
                 except FileNotFoundError as e:
                     last_md5 = None
+                    logging.debug(f"Unable to load {latest_file} - setting last md5 to none")
             else:
                 last_md5 = None
+                logging.debug(f"{latest_file} does not exist - setting last md5 to none")
 
             # Download new image and generate MD5
             image = requests.get(image_link, allow_redirects=True)
             image_md5 = hashlib.md5(image.content).hexdigest()
+            logging.info(f"md5sum of downloaded image: {image_md5}")
+
+            if image.status_code != 200:
+                logging.error(f"Unable to download image: {image_link}")
+                exit(1)
 
             # Since NASA timezone and my timezone is not aligned and the file from yesterday might still
             # be the "latest" file on NASA's website, then we check if the current file's MD5 sum matches
             # what we downloaded. If it does, then we don't do anything as we already have the latest file.
             if image_md5 != last_md5:
+                logging.debug(f"MD5 hashes did not math, {image_name} is a new file")
 
                 # Save downloaded image
+                logging.debug(f"Writing {destination_file}")
                 with open(destination_file, "wb") as f:
                     f.write(image.content)
 
                 # Unlink last image
                 if path.islink(latest_file):
                     os.unlink(latest_file)
+                    logging.debug(f"{latest_file} deleted")
 
                 # Relink downloaded image
                 if not path.islink(latest_file):
                     os.symlink(destination_file, latest_file)
+                    logging.debug(f"{destination_file} symlinked to {latest_file}")
 
-                print(f"Downloaded image: {destination_file}")
+                logging.critical(f"Downloaded image: {image_name}")
+            else:
+                logging.error(f"The image '{image_name}' has already been downloaded previously")
 
         # End of execution - break out of for loop
-        break
+        exit(0)
